I have forgotten about this blog post for almost a year. I got reminded of it while writing the following reply to someone who said lucky coincidences that saved characters were the biggest defining factor whether they liked a movie or not, in retrospect. (http://news.ycombinator.com/item?id=4087498)
------------------------

A coincidence is just something which happened without a lead-on. An outcome surprising, therefore perceived as improbable (perception bias at play here).

Anything that happens in time has an outcome and things that follow it in logical order. However, we think differently if we flip the time axis, and look at the sequence of events leading onto a final event, rather than if we look at the sequence of events started by a seminal event.

The simple way to overcome this?

Put your coincidence in. Then, start going back in time, dropping smaller and smaller clues and building premise for the "coincidence" to happen. Finally, it's not a coincidence any more.

The difference between "coincidence" and "culmination" is fairly menial work. I am surprised more authors don't do this, and sometimes wonder if they don't include a build-up just to be intentionally cheesy.

If you want to have fun experimenting with the reversal of time in logical reasoning, please try a game I have invented, called Neutrino. You can find a description at http://cheater.posterous.com/neutrino (let me know if there are any problems). I have come up with it autumn last year and wrote a blog post about it, but forgot to upload it (or maybe there was some more work to do that I forgot about completely...) - you get an upvote for several reasons, one of them is that you reminded me that I should put it up! Thanks!